[BTC-109] - Implement some basic layers and masks so projectiles can collide with level
Added some layers and masks to make the collision be more precise and we destroy the projectile after collision
Added some primitives to the world environment

TIP: issue areas don't collide with static bodies
If you also found yourself here and are using Jolt3d. The answer is in the advanced project settings of Jolt3D. You will need to enable "Areas detect static bodies". This is not enabled by default.

Resolves #109